requirements.txt
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-keyvault-secrets==4.8.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
flask-paginate==2024.4.12
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
msal==1.28.0
msal-extensions==1.1.0
packaging==24.0
portalocker==2.8.2
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
